infrastructure/azure-functions/document_processor/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
openai>=1.0.0
python-dotenv>=1.0.1
